refactor(gym): extract checkpoint dataclasses to gym/checkpoint.py (#115, step 1/N)
First step in splitting MicroPlanRunner (2960 LOC). Pure code move +
re-exports, zero behavior change.

Moved to new module mantis_agent.gym.checkpoint:
- StepResult       (round-trips through checkpoint JSON via _PERSISTED)
- RunCheckpoint    (cross-session resume state, all 25 fields)
- PauseRequested   (host tool exception for pause requests)
- PauseState       (serializable paused-runner snapshot)
- RunnerResult     (public return type from run()/resume())
- REVERSE_ACTIONS  (per-step recovery key presses)

Why these go first:
- Pure data with no execution-loop dependency.
- Smallest blast radius (callers only use these for type hints + load/save).
- Sets the import pattern for the next 5 extractions
  (CostMeter, ToolChannel, BrowserState, ListingDedup, Executor).

Backward compat:
- Every name still importable from mantis_agent.gym.micro_runner.
- mantis_agent.__getattr__ shim untouched (still resolves through
  micro_runner, which re-exports the new module).
- The 7 internal + external import sites work unchanged
  (modal_cua_server, baseten_server/runtime, graph/learner,
  test_micro_runner_hooks, test_form_intents, test_private_seller_filter,
  test_prompt_versions).

micro_runner.py: 2960 → 2866 LOC (-94)
new checkpoint.py: 220 LOC

Tests:
- 6 new unit tests in test_checkpoint_module.py covering canonical-vs-
  legacy import identity, round-trip persistence with #127 fields,
  PauseRequested extras, observability-extra exclusion from to_dict().
- 54 existing tests pass across test_form_intents, test_prompt_versions,
  test_no_orphan_submodules, test_brain_protocol, test_checkpoint_module.
- ruff clean.
